+++ docs/Changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## unreleased
+* feat: [`run`][mininterface.run] add_config flag
+
 ## 1.1.4 (2025-10-13)
 * enh: Python 3.14 compatible